requirements.txt
python-dotenv==0.14.0
pyautogui
customtkinter